requirements-test.txt
# Testing framework requirements for Llama Mapper multi-service testing
# This file contains all dependencies needed for comprehensive testing across:
# - Core Mapper Service
# - Detector Orchestration Service  
# - Analysis Service

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# Performance testing
pytest-benchmark>=4.0.0
locust>=2.17.0
memory-profiler>=0.61.0
py-spy>=0.3.14

# Security testing
pytest-security>=0.3.0
bandit>=1.7.5
safety>=2.3.5
semgrep>=1.45.0

# Chaos engineering
chaostoolkit>=1.16.0
chaostoolkit-kubernetes>=0.27.0
chaostoolkit-prometheus>=0.8.0

# Database and service clients
asyncpg>=0.28.0
redis>=5.0.1
clickhouse-driver>=0.2.6
psycopg2-binary>=2.9.7

# HTTP clients and mocking
httpx>=0.25.0
requests>=2.31.0
responses>=0.23.3
pytest-httpserver>=1.0.8

# Container and environment management
docker>=6.1.3
docker-compose>=1.29.2

# Data generation and validation
faker>=19.6.2
hypothesis>=6.87.0
factory-boy>=3.3.0

# Data processing and analysis
pandas>=2.1.1
numpy>=1.25.2
structlog>=23.1.0

# Monitoring and observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-asyncpg>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tenacity>=8.2.3
click>=8.1.7

# Cryptography for security testing
cryptography>=41.0.4
pyjwt>=2.8.0

# Static analysis and code quality
mypy>=1.5.1
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
pylint>=2.17.7

# Coverage and mutation testing
coverage>=7.3.2
mutmut>=2.4.3
cosmic-ray>=8.3.6

# Contract testing
pact-python>=2.0.0

# JSON schema validation
jsonschema>=4.19.1
cerberus>=1.3.5

# Test reporting and visualization
allure-pytest>=2.13.2
pytest-clarity>=1.0.1
pytest-sugar>=0.9.7

# Parallel testing and distribution
pytest-parallel>=0.1.1
pytest-testmon>=2.0.0

# Development utilities (for test development)
ipython>=8.15.0
jupyter>=1.0.0
matplotlib>=3.7.2
seaborn>=0.12.2

# Mock services and fixtures
wiremock>=2.2.0
testcontainers>=3.7.1

# Load testing visualization
locust-plugins>=4.3.0

# API testing
tavern>=2.4.1

# Property-based testing
hypothesis-auto>=1.1.4

# Test data management
mimesis>=11.1.0
